First commit
After a lot of fighting with the serial port and python, trying three
languages and about ten libraries, including native Windows serial
interface, I had to go back to Python, which was the only language I
managed to get working with serial port reliably... although slowly. But
it is still much faster than the original program.